+++ WorldCities/WorldCitiesAPI/Data/Models/ApplicationUser.cs
@@ -8,7 +8,7 @@ namespace WorldCitiesAPI.Data.Models
   /// We dont neet to implement anything here; we'll just extend the IdentityUser base class, which already
   /// contains everything we need for now. 
   /// </summary>
-  public class ApplicationUser: IdentityUser
+  public class ApplicationUser : IdentityUser
   {
 
   }